ml/requirements.txt
fastapi
uvicorn
pydantic
python-multipart
pandas
scikit-learn
numpy
joblib
google-cloud-storage
